requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
colorama==0.4.1
Flask==1.0.3
idna==2.8
importlib-metadata==0.17
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
more-itertools==7.0.0
mysql==0.0.2
mysqlclient==1.4.2.post1
numpy==1.16.4
packaging==19.0
pandas==0.24.2
Pillow==6.0.0
pluggy==0.12.0
py==1.8.0
pyaml==19.4.1
pyparsing==2.4.0
pytest==4.6.2
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1.1
requests==2.22.0
six==1.12.0
SQLAlchemy==1.3.4
urllib3==1.25.3
wcwidth==0.1.7
Werkzeug==0.15.4
zipp==0.5.1
